+++ JobsUkraine/src/main/java/ua/com/jobsukraine/service/CandidateService.java
@@ -28,5 +28,5 @@ public interface CandidateService {
 	
 	void acceptVacancy(Candidate candidate, Vacancy vacancy);
 	
-	List<Candidate> getRandomBestCandidates(int amount, int minRating);
+	List<Candidate> getRandomBestCandidates(int amount, double minRating);
 }
